created login, refined routes, edit models

+++ src/routes/users.ts
@@ -1,30 +1,28 @@
 import express, { Request, Response } from 'express';
-import { userSchema, UserAcessLevel } from '../models/userSchema';
+import { userSchema, UserAccessLevel } from '../models/userSchema';
 import { errorHandler } from '../functions/errorHandler';
 import validator from 'validator';
-import jwt from 'jsonwebtoken';
-import mongoose, { ObjectId } from 'mongoose';
+import jwt, { JwtPayload } from 'jsonwebtoken';
+import bcrypt from 'bcrypt';
 
 const router = express.Router();
 
 router.get("/users", async (req: Request, res: Response) => {
     /* 
     #swagger.tags = ['Users']
+    #swagger.parameters['token'] = { in: "query", name: "token", required: false }
     #swagger.responses[200] = { description: "OK", schema: { $ref: "#/components/schemas/UserData" } }    
     */
 
     try {
+        const token = jwt.decode(String(req.query.token)) as JwtPayload;
+        const userLogged = await userSchema.findOne({ _id: token }).select("+accessLevel");
 
-        // const token = req.params.token;
+        if (userLogged.acessLevel !== UserAccessLevel.admin) {
+            throw new Error("Unauthorized!");
+        }
     
-        // const decodedToken = jwt.decode(token);
-        // const user = await userSchema.findOne({ _id: decodedToken }).select("+acessLevel");
-
-        // if (user.acessLevel !== UserAcessLevel.admin) {
-        //     throw new Error("Unauthorized!");
-        // }
-    
-        const users = await userSchema.find().select("+token");
+        const users = await userSchema.find().select(["+acessLevel", "+token"]);
         res.status(200).json(users);
 
     } catch (error) {
@@ -36,14 +34,23 @@ router.get("/users", async (req: Request, res: Response) => {
 router.get("/users/:id", async (req: Request, res: Response) => {
     /* 
     #swagger.tags = ['Users']
+    #swagger.parameters['token'] = { in: "query", name: "token", required: false }
     #swagger.responses[200] = { description: "OK", schema: { $ref: "#/components/schemas/UserData" } }
     #swagger.responses[404] = { description: "Not Found", schema: { $ref: "#/components/schemas/ErrorMessage" } }
     */
 
     const id = req.params.id;
     try {
-        const book = await userSchema.find({ _id: id });
-        res.status(200).json(book);
+
+        const token = jwt.decode(String(req.query.token)) as JwtPayload;
+        const userLogged = await userSchema.findOne({ _id: token }).select("+accessLevel");
+
+        if (userLogged.accessLevel !== UserAccessLevel.admin && String(userLogged._id) !== id) {
+            throw new Error("Unauthorized!");
+        }
+
+        const user = await userSchema.find({ _id: id });
+        res.status(200).json(user);
 
     } catch (error) {
         if (String(error).includes("_id")) {
@@ -57,30 +64,42 @@ router.get("/users/:id", async (req: Request, res: Response) => {
 router.post("/users", async (req: Request, res: Response) => {
     /* 
     #swagger.tags = ['Users']
+    #swagger.parameters['token'] = { in: "query", name: "token", required: false }
     #swagger.responses[200] = { description: "OK", schema: { $ref: "#/components/schemas/UserData" } }
     #swagger.responses[400] = { description: "Invalid Body", schema: { $ref: "#/components/schemas/ErrorMessage" } }
     #swagger.requestBody = { content: { "application/json": { schema: { $ref: "#/components/schemas/User" } } } }
     */
 
-    const { name, email, password, acessLevel } = req.body;
-
     try {
+
+        const token = jwt.decode(String(req.query.token)) as JwtPayload;
+        const userLogged = await userSchema.findOne({ _id: token }).select("+accessLevel");
+
+        if (userLogged.accessLevel !== UserAccessLevel.admin) {
+            throw new Error("Unauthorized!");
+        }
+
+        const { name, email, password, accessLevel } = req.body;
+
+
         if (!validator.isEmail(email)) {
             throw new Error("Invalid Email!");
         }
 
-        const dbResponse = await userSchema.create({ name, email, password, acessLevel });
-        //const user = await userSchema.findOne({ email });
+        const hashRounds = 10;
+        const hashPassword = await bcrypt.hash(password, hashRounds);
 
-        //const token = jwt.sign(user._id, process.env.JWT_SECRET!);
-
-        //await userSchema.updateOne({ email }, { token });
-
+        const dbResponse = await userSchema.create({ name, email, password: hashPassword, accessLevel });
+        
+        const genToken = jwt.sign(String(dbResponse._id), process.env.JWT_SECRET!);
 
         res.status(200).json({
             status: "OK",
             message: `User has been created!`,
-            response: dbResponse
+            response: {
+                token: genToken,
+                user: dbResponse
+            }
         });
 
     } catch(error) {
@@ -91,6 +110,7 @@ router.post("/users", async (req: Request, res: Response) => {
 router.put("/users/:id", async (req: Request, res: Response) => {
     /* 
     #swagger.tags = ['Users']
+    #swagger.parameters['token'] = { in: "query", name: "token", required: false }
     #swagger.responses[200] = { description: "OK", schema: { $ref: "#/components/schemas/UserData" } }
     #swagger.responses[400] = { description: "Invalid Body", schema: { $ref: "#/components/schemas/ErrorMessage" } }
     #swagger.responses[404] = { description: "Not Found", schema: { $ref: "#/components/schemas/ErrorMessage" } }
@@ -98,10 +118,22 @@ router.put("/users/:id", async (req: Request, res: Response) => {
     */
 
     const id = req.params.id;
-    const { name, password, acessLevel } = req.body;
 
     try {
-        const dbResponse = await userSchema.updateOne({ _id: id }, { name, password, acessLevel });
+
+        const token = jwt.decode(String(req.query.token)) as JwtPayload;
+        const userLogged = await userSchema.findOne({ _id: token }).select(["+password", "+accessLevel"]);
+
+        if (userLogged.accessLevel !== UserAccessLevel.admin) {
+            throw new Error("Unauthorized!");
+        }
+
+        const { name, password, accessLevel } = req.body;
+
+        const hashRounds = 10;
+        const hashPassword = await bcrypt.hash(password, hashRounds);
+
+        const dbResponse = await userSchema.updateOne({ _id: id }, { name, password: hashPassword, accessLevel });
 
         if (dbResponse.modifiedCount > 0) {
             const user = await userSchema.findOne({ _id: id });
@@ -122,62 +154,6 @@ router.put("/users/:id", async (req: Request, res: Response) => {
     }
 });
 
-router.put("/users/:id/regentoken", async (req: Request, res: Response) => {
-    /* 
-    #swagger.tags = ['Users']
-    #swagger.responses[200] = { description: "OK", schema: { $ref: "#/components/schemas/UserData" } }
-    #swagger.responses[404] = { description: "Not Found", schema: { $ref: "#/components/schemas/ErrorMessage" } }
-    */
-   
-    const id = req.params.id;
-    try {
-        const maxAttempts = 10;
-        let attempts = 1;
-        let sucessRegen = false;
-        let newId = "";
-
-        while(!sucessRegen) {
-            newId = String(new mongoose.Types.ObjectId());
-            
-            const user = await userSchema.findOne({ _id: newId });
-            if (user) {
-                if (attempts >= maxAttempts) {
-                    throw new Error("Unable to regenerate your token, please try again");
-                }
-
-                attempts++;
-                continue;
-            }
-
-            sucessRegen = true;
-        }
-
-        const newToken = jwt.sign(newId, process.env.JWT_SECRET!);
-        const dbResponse = await userSchema.updateOne({ _id: id }, { _id: newId, token: newToken });
-        
-        if (dbResponse.modifiedCount > 0) {
-            const user = await userSchema.findOne({ _id: newId });
-
-            res.status(200).json({
-                status: "OK",
-                message: `User (#${id}) has his token regenerated!`,
-                response: {
-                    token: user
-                }
-            });
-        }
-
-    } catch(error) {
-        // if (String(error).includes("_id")) {
-        //     return errorHandler(res, new Error("User not found!"), { errorTitle: "Not Found", errorStatusCode: 404 });
-        // }
-
-        errorHandler(res, error);
-    }
-
-
-});
-
 router.delete("/users/:id", async (req: Request, res: Response) => {
     /* 
     #swagger.tags = ['Users']
@@ -188,19 +164,18 @@ router.delete("/users/:id", async (req: Request, res: Response) => {
     */
 
     const id = req.params.id;
-    //const token = req.params.token;
     try {
 
-        // const decodedToken = jwt.decode(token);
-        // const user = await userSchema.findOne({ _id: decodedToken }).select("+acessLevel");
+        const token = jwt.decode(String(req.query.token)) as JwtPayload;
+        const userLogged = await userSchema.findOne({ _id: token }).select("+accessLevel");
 
-        // if (user.acessLevel !== UserAcessLevel.admin) {
-        //     throw new Error("Unauthorized!");
-        // }
+        if (userLogged.accessLevel !== UserAccessLevel.admin) {
+            throw new Error("Unauthorized!");
+        }
 
-        // if (id === decodedToken) {
-        //     throw new Error("You can't delete your self!");
-        // }
+        if (userLogged._id === id) {
+            throw new Error("You can't delete your self!");
+        }
 
         const dbResponse = await userSchema.deleteOne({ _id: id });
